modularisation+commentaires POO
