+++ L2/S2/POO/Projet/carte.h
@@ -2,6 +2,10 @@
 #define carte_h
 
 #include <iostream>
+#include <sstream>
+#include <string>
+
+//TODO COMMENTAIRES CPP & H
 
 class Carte
 {
@@ -16,8 +20,8 @@ class Carte
     int getValeur() const;
     int getFamille() const;
     Carte& operator = (const Carte& c);
-    void afficherCouleur();
-    void afficherValeur();
+    std::string afficherFamille() const ;
+    std::string afficherValeur() const;
     friend std::ostream& operator << (std::ostream& os, const Carte& c);
 };
